Diffie Hellman  (#131)
* The layout of the left buttons changed to GridLayout

* Entire Layout changed to GridLayout

* Layout of the groups Alice and Bob changed

* Arrows are variable now

* GUI improvements

The GUI has looks like before, but is variable now.
Two bugs still exist. First, when the animation comes, the arrows in the
canvasExchange disappear. Second, a key should still be faded in - does
not happen yet.

* Changes the position of Button Random Curve #33 Nr.3

* various changes

Question mark characters removed in the dialogues.
Improved user input validation in PublicParametersDialog.
Warning in case of incorrect input

* No help icons anymore / Comments deleted

* faulty state: see stdout output showing the getBounds for gd_stDescription being zero

* correct calculation of title label size, see it's javadoc

* mostly smaller changes

Some texts changed.
The order of G is displayed in the "Select Secret" dialog if the curve
size is large.
The animation, transitioning the binary code from Alice to Bob, adapts
to the size of the canvas.
key.png now has an alpha channel -> is transparent.
Display location of the key changed slightly.
Addresses #117 and #33. In #33 all issues are solved.
Removed redundant redraw() commands.
Problem: The animation stops on the way from Bob to Alice, was
previously so.

* Key icon will be disposed at reset.

* The groups alice/bob do not resize when opening / closing other windows.

* The key icon is displayed under the groups Alice and Bob

* Spelling mistakes fixed / Secret Key Wizard page changed

The order of G is no longer displayed in the description part of the
dialog, but above the entry field for the secret on large curves

* Line breaks replaced by blank characters in textCurve.

* Fixes #33 Nr.5

The value of G gets changed when changing the curve size from large to
small.
Removed reused GridData.
Replaced <sup> with ^ and removed </sup> in textCurve.
Added blank spaces to textCurve.