GUI fixes in Certificate Verification Plugin + ElGamal GUI fixes (#164)
* Button enter text renamed to enter plaintext or enter ciphertext

* various GUI improvements

Added the c to the setFocus() to enable direct scrolling.
Removed the unnecessary TabItem.
Renamed many variables to fit the java naming conventions.
Fixed the label that display whether a verification was successfull or
not.
Removed paintListerners from load Buttons.
Added code to the selectionListener of the pki button to open the pki
plugin.

* Fixed title internationalization

* Increased the widthHint of the description so that no vertical scrollbar

is shown on 1080p monitors.

* Text changes

* Commented out the reset button, the pki plugin button, the signature

verification plugin.

* Duplicate entries for the restart and help removed.

* Improved the input validation.

No exceptions are thrown anymore.
The BigIntegers r and s are only set if the Input has the form of a
tuple. Example: (ab3,342)

* Removed duplicate entry for restart in the plugin menu bar.

* Fixed a malformed unicode character.

* Button added for plain text input + various other changes.

Seperated fields for signature and plaintext input in verification tab.
For that i added the PlaintextForVerificationWizard and the
EnterPlainTextForVerification wizard page.
Removed NewTextEntryWizard that i added a commit before.
Fixed broken externalized strings.
Improved input validation for the signature input
Removed redundant action object in the signature input wizard.
Some texts changed.
Varibles renamed to have the same names as in the online help.
Several minor changes.

* Removed unnecessary actions object.

Fixed a spelling error.
In the NewKeypairDialog no exception will be thrown if you enter b
earlier.

* Fixes a problem that caused no result to be displayed when trying to
validate a signature if plain text was not typed.

* The title is adapted to the current action.

* Added a copy to clipboard button right near the step button.

* Added a label left near the Start button for the stepwise calculation,
so that it is more clear for what the button is.

* Added a line break to a dialog description.

* Added screenshots to the ElGamal online help.

Different pictures in de and en.