Signature visualization – UI/UX improvements, text/typo corrections, online help fixed (#169)
* Test commit

* GUI Improvement of the plugin (AbsoluteLayout -> GridLayout).

* Undo test commit

* Further GUI improvement/bug fixing

* Added description at the top and improved German translation

* GUI Improvements of the Signature demonstration Plug-In

- Changed the Layout to GridLayout in the main composite
- Changed the Layout to GridLayout in all wizards
- Improved usability and readability of the GUI components
- Layout is responsive to the display size, now
- Text is not cropped on High-DPI displays, anymore

* Small fixes for extremely high DPI setting/scaling

* Signatur visualization – Several improvements (mainly GUI)

- Corrected/optimized text at several different positions in the plugin
itself and the online help.
- Input focus is initially set now in all dialog windows.
- I added a tooltip that shows the message / the beginning of the
message that the user entered and has a context menu so that the user
can copy the tooltip into the clipboard.
- I added a tooltip that shows the name of the used key for signing the
message. 
- When re-opening a dialog window, the user now sees his formerly
entered input values.
- When re-opening the „Choose input“ dialog window in a later step: If
the user does not change the input message, no reset takes place. Only
if the user changes the input message, a already computed hash and
signature message get discarded.
- I removed the former limit of 10MB for input files. For replacement, I
implemented a dynamic limitation. It allows for much bigger files and
depends on the amount of heap space that JCT is able to acquire.
- Before this update, hex values have been computed for the whole input
file/message. Since the hex representation needs a very big amount of
memory, and since only a small fraction of the file/message gets
displayed in hex representation to the user, I limited this computation
to the first 1024 bytes of the input file/message.
- All comob boxes are now initially filled with the first possible
value.
- I implemented a context menu for the tables / text areas in the last
dialog window („Show generated signature“), so that the user can copy
one row/the complete table into the clipboard.
- I implemented a button „Open in hex editor“ in the last dialog window
(„Show generated signature“). It is only accessible if the message/file
was saved before, using the „Save“ button.
- I made the „Show generated signature“-dialog window adjustable in
size. The tables expand in a responsive way when stretching it in width. 
- I repaired some links in the online help.