Code cleanup Fixed description part
removed border
fixed size
Removed fix cell width from the composite
Set the size to be automatically calculated with

scroll.setMinSize(arc4.computeSize(SWT.DEFAULT, SWT.DEFAULT));
Changed the GridData of all cells to grab the vertical space.

Switched from MouseListener to SelctionListener
Added a border around the input text so it looks more like you can enter

a value there.
Various changes

- added line breaks \n to long tooltip texts so they don't cover the
whole display.
- replaced SWT.SHOADOW_IN style constants with SWT.NONE in group
constructors to avoid later issues. They don't have an effect on Win 7.
- removed fix pixel specifications for several Controls, especially the
size of the scrolledComposite and the vector labels. Replaced them with
the default size that will be computed during a layout() call.

Small issue:
At the bottom of the page is a little empty space. This is due to the
widthHint of the description Text. The width hint is set to cause the
Text to wrap.
Removed magenta Background :D