Fixed red possibilities. Everything is black now. Various changes
- Added a restart icon in the top right corner of the view to reset the
current view.
- Removed some system.out.print
Fix for long running job that can have a reference to a control that my

be disposed when hitting the reset button.
Given values are printed bold and unchangeable
The reset button is renamed to restart and it resets the sudoku to its

initial state.
Added a restart button to the hex sudoku and killer sudoku tab.