Added automatic save/reload function.
Values that you enter will be saved when exit the wizard. It is saved in
both cases, when you exit the dialog with the X or the cancel button and
when you hit finish.
When you restart the wizard the values are set on startup.
Added a reset button.
Replaced umlaut with unicode char
Output option is now also saved.
Added myself as author for the load/save method
Fixed a typo.