Fix for: commands in toolbar not enabled
The commands in the toolbar are not enabled when a state changes. So the
redo/ undo/ sharkMeal/ reset-Commands only worked in the
menuContribution menu:, not in the menuContribution toolbar:. The
Problem was, that <activeWhen></aviveWhen> is used to show or hide
commands and it does not work with the menuContribution toolbar:. The
solution was to change the <activeWhen> to <enabledWhen>. This is the
way the greying ungreying should be done.