Fixes a problem that causes all TLS 1.3 connections to fail.
The previous implementation listed the flexiprovider crypto-providers as
the first providers to be used. However, TLS 1.3 connections sometimes
require very long keys, which the flexiprovider no longer supports. It
is already a few years old.
Now the SUN crypto providers are used first. These are the standard and
support long keys.